pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "audioman"
authors = [
  { name="ego-lay-atman-bay" },
]
# version = "0.0.1"
description = "Python library for manipulating audio files. This can be used for editing audio, or tagging audio."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
  "filetype",
  "mutagen",
  "numpy",
  "Pillow",
  "setuptools",
  "soundfile",
  "charset_normalizer",
  "pydub",
  "audioop-lts", # to fix pydub in python 3.13
  "pathvalidate",
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/ego-lay-atman-bay/audioman/"
"Bug Tracker" = "https://github.com/ego-lay-atman-bay/audioman/issues/"
"Documentation" = "https://ego-lay-atman-bay.github.io/audioman/"

[tool.setuptools.dynamic]
version = { attr = "audioman.__version__" }
